requirements.txt
crewai>=0.28.0
groq>=0.4.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-groq>=0.1.0